move logic outside gui to shape editor folder
