+++ waveform_editor/gui/shape_editor/plasma_shape.py
@@ -3,7 +3,7 @@ import panel as pn
 import param
 from panel.viewable import Viewer
 
-from waveform_editor.gui.shape_editor.logic.plasma_shape_calc import (
+from waveform_editor.shape_editor.plasma_shape_calc import (
     Gap,
     compute_outline_from_params,
     update_outline_from_gaps,
